+++ zmeventnotification/zmes_hook_helpers/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "6.1.12"
+__version__ = "6.1.15"
 VERSION=__version__